requirements.txt
mcp
paho-mqtt
python-dotenv
